1. request change to application/json and method only support post 2. roleid limit into integer 3. acct -> account 4. remove ud in protobuf.MessageHeader 5. auto generate gamesrv's config 6. logger.log->logger.logf 7. bugfix: lkcp.c segment fault

+++ gamesrv/src/gg/base/profile.lua
@@ -35,7 +35,7 @@ function profile.stat(record,name,onerror,func,...)
 		if profile.threshold and time > profile.threshold then
 			cost.overloadcnt = cost.overloadcnt + 1
 			if profile.log_overload then
-				logger.log("info","profile","op=overload,name=%s,time=%ss",name,time)
+				logger.logf("info","profile","op=overload,name=%s,time=%ss",name,time)
 			end
 		end
 	end
